requirements.txt
asgiref>=3.8
azure-identity>=1.17
azure-keyvault-secrets>=4.8
cryptography>=42.0
Django>=5.0,<6.0
gunicorn>=22.0
psycopg[binary]>=3.2
social-auth-app-django>=5.4
sqlparse>=0.5
